refactor: use layers within category route handlers
